chore: move components to component package
