+++ app/src/main/java/org/sopt/official/feature/attendance/compose/component/MidtermAttendanceCard.kt
@@ -1,4 +1,4 @@
-package org.sopt.official.feature.attendance.compose
+package org.sopt.official.feature.attendance.compose.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
